+++ Source/AccelByteUe4Sdk/Public/Api/AccelByteSessionApi.h
@@ -402,6 +402,17 @@ public:
 		, THandler<FJsonObjectWrapper> const& OnSuccess
 		, FErrorHandler const& OnError);
 
+	/**
+	 * @brief Query recently met users in a game session.
+	 *
+	 * @param OnSuccess This will be called if the operation succeeded.
+	 * @param OnError This will be called if the operation failed.
+	 * @param Limit Number of recent players to request, maximum value is 200.
+	 */
+	void GetRecentPlayers(THandler<FAccelByteModelsV2SessionRecentPlayers> const& OnSuccess
+		, FErrorHandler const& OnError
+		, const int32 Limit = 20);
+
 private:
 	Session() = delete;
 	Session(Session const&) = delete;